v1 of completely rebuilt tokenizer/parser
This work leans heavily on the Java reference implementation of
Zinc reading/writing i found on BitBucket from Brian Frank:
https://bitbucket.org/brianfrank/haystack-java

Performance is worse, but probably because parsing quality and
comprehensiveness is significantly up. We're still about 40-50x
faster than hszinc.

There are a few failing tests right now, but it's for a relatively
minor issue of an int being round-tripped into a float.